Checkpoint: Distributed Crawler intermediate progress.
Project is in a working-ish state, however, a number of items need to
be sorted out and cleaned up:

- Finish disentangling the crawler manager thing from actual crawler
  worker component.

- Differentiation between local error conditions which can be simply
  logged (or saved) alongside the crawl snapshot messages and then
  ignored vs. fatal error conditions which must cause the crawler loop
  to stop or something else special ahppen.

  Examples of unimportant erros:
    * Problem checking out package src code
    * Problem obtaining package RepoRoot
    * Malformed package structure

  Examples of errors which are significant to the loop:
    * Error saving a package or to-crawl entry
    * Disk full

- Come up with a better name than "CrawlerCoordinator".  I'd thought of
  a good one last night but didn't write it down because I was certain
  I'd remember it.  Oops.

TODO: Add a monitor and require that the disk where the DB is stored
always has at least X GB free, where X is based on a multiple of the
Bolt database file.  This is to ensure safety that things don't get into
a state where data cannot be written to the DB or even worse it gets
corrupt.  Remember that DB size may grow non-linearly (need to double
check this, but this is what I recall observing).
(Also now noted in README TODOs list.)